feat: create screen for select state management
