+++ lib/presenter/pages/covid/pages/cases/cases_page.dart
@@ -1,68 +1,30 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
-import 'cases_state.dart';
-import 'widgets/search_bar_widget.dart';
-import 'widgets/case_card_widget.dart';
-import '../../../../widgets/load_notifier/load_notifier_builder_widget.dart';
-import '../../../../../domain/entities/covid/case_entity.dart';
-
-class CasesPage extends StatefulWidget {
-  @override
-  _CasesPageState createState() => _CasesPageState();
-}
-
-class _CasesPageState extends ModularState<CasesPage, CasesState> {
-  final searchController = TextEditingController();
+class CasesPage extends StatelessWidget {
+  const CasesPage({
+    Key? key,
+  }) : super(
+          key: key,
+        );
+  
+  static const Map<String, String> items = {
+    'ValueNotifier': '/covid/cases/value-notifier',
+  };
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Colors.blueGrey[900],
       appBar: AppBar(
         title: Text('Covid-19 - Cases'),
-        backgroundColor: Colors.blueGrey[900],
-        actions: [
-          ValueListenableBuilder<bool>(
-            valueListenable: controller.searchBarVisibleListenable,
-            builder: (context, showSearchBar, widget) => IconButton(
-              onPressed: () =>
-                  controller.searchBarVisibleListenable.value = !showSearchBar,
-              icon: Icon(showSearchBar ? Icons.close : Icons.search),
-            ),
-          ),
-        ],
       ),
-      body: Column(
-        children: [
-          ValueListenableBuilder<bool>(
-            valueListenable: controller.searchBarVisibleListenable,
-            builder: (context, showSearchBar, widget) => showSearchBar
-                ? Padding(
-                    padding: const EdgeInsets.all(8.0),
-                    child: SearchBarWidget(
-                      onChanged: (String value) =>
-                          controller.textSearchListenable.value = value,
-                    ),
-                  )
-                : Container(),
-          ),
-          Expanded(
-            child: LoadNotifierBuilderWidget<List<CaseEntity>?>(
-              valueListenable: controller.casesListenable,
-              onLoading: (context) => Center(
-                child: CircularProgressIndicator(),
-              ),
-              onError: (context, error) => Center(
-                child: Text('Ocorreu um erro ao carregar!'),
-              ),
-              onDone: (context, cases) => ListView.builder(
-                itemCount: cases!.length,
-                itemBuilder: (context, index) => CaseCardWidget(cases[index]),
-              ),
-            ),
-          ),
-        ],
+      body: ListView.builder(
+        itemBuilder: (_, index) => ListTile(
+          title: Text('${items.keys.elementAt(index)}'),
+          trailing: Icon(Icons.arrow_forward),
+          onTap: () => Modular.to.pushNamed('${items.values.elementAt(index)}'),
+        ),
+        itemCount: items.length,
       ),
     );
   }
