+++ lib/presenter/pages/covid/pages/cases/value_notifier/cases_state.dart
@@ -1,9 +1,9 @@
-import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
 
-import '../../../../../domain/entities/covid/case_entity.dart';
-import '../../../../../domain/usecases/covid/search_cases_usecase.dart';
-import '../../../../../domain/usecases/covid/get_cases_usecase.dart';
-import '../../../../widgets/load_notifier/load_notifier_listenable.dart';
+import '../../../../../../domain/entities/covid/case_entity.dart';
+import '../../../../../../domain/usecases/covid/search_cases_usecase.dart';
+import '../../../../../../domain/usecases/covid/get_cases_usecase.dart';
+import '../../../../../widgets/load_notifier/load_notifier_listenable.dart';
 
 class CasesState extends ChangeNotifier {
   final _casesListenable = LoadNotifierListenable<List<CaseEntity>?>(null);
